Refactor AI contract tests for improved structure and clarity
