+++ tools/release/tests/changelog/test_release_context.py
@@ -4,7 +4,9 @@ import unittest
 from unittest.mock import patch
 
 from tools.release.changelog.context_builder import build_release_context
+from tools.release.changelog.git_collect import get_commits_since_tag, get_previous_release_tag_before
 from tools.release.changelog.models import CommitInfo
+from tools.release.version.models import Version
 
 
 class ContextBuilderTests(unittest.TestCase):
@@ -119,6 +121,59 @@ class ContextBuilderTests(unittest.TestCase):
         ):
             context = build_release_context(version="0.34.2", repo_root=".")
         self.assertEqual([commit.sha for commit in context.commits], ["newest", "older"])
+
+
+
+class GitCollectTests(unittest.TestCase):
+    def test_empty_body_commit_is_not_dropped(self) -> None:
+        git_log = "6cf9a156eae61fd9ba2c2204ee9c73c2832bd14b\x1fUpdate docs\x1f\x1e"
+
+        with (
+            patch("tools.release.changelog.git_collect._run_git", return_value=git_log),
+            patch(
+                "tools.release.changelog.git_collect.get_commit_file_summary",
+                return_value=(["README.md"], 2, 1),
+            ),
+        ):
+            commits = get_commits_since_tag(".", "v0.30.5")
+
+        self.assertEqual(len(commits), 1)
+        self.assertEqual(commits[0].sha, "6cf9a156eae61fd9ba2c2204ee9c73c2832bd14b")
+        self.assertEqual(commits[0].subject, "Update docs")
+
+    def test_metadata_only_packaging_release_commit_gets_non_meta_categories(self) -> None:
+        git_log = (
+            "6cf9a156eae61fd9ba2c2204ee9c73c2832bd14b\x1f"
+            "update packaging documentation; clarify Debian build orchestration and release tooling structure"
+            "\x1f\x1e"
+        )
+
+        with (
+            patch("tools.release.changelog.git_collect._run_git", return_value=git_log),
+            patch(
+                "tools.release.changelog.git_collect.get_commit_file_summary",
+                return_value=(["README.md"], 3, 0),
+            ),
+        ):
+            commits = get_commits_since_tag(".", "v0.30.5")
+
+        self.assertEqual(len(commits), 1)
+        self.assertIn("meta", commits[0].categories)
+        self.assertIn("debian", commits[0].categories)
+        self.assertIn("tools", commits[0].categories)
+
+    def test_previous_release_tag_before_base_prefers_nearest_lower_tag(self) -> None:
+        tags = "\n".join(["v0.33.0", "v0.32.1", "v0.34.1", "v0.35.0"])
+        with patch("tools.release.changelog.git_collect._run_git", return_value=tags):
+            resolved = get_previous_release_tag_before(".", Version(0, 34, 0))
+        self.assertEqual(resolved, "v0.33.0")
+
+    def test_previous_release_tag_before_handles_sparse_history(self) -> None:
+        tags = "\n".join(["v0.33.0", "v0.58.7", "v0.58.8"])
+        with patch("tools.release.changelog.git_collect._run_git", return_value=tags):
+            resolved = get_previous_release_tag_before(".", Version(0, 58, 0))
+        self.assertEqual(resolved, "v0.33.0")
+
 
 
 if __name__ == "__main__":
